chore: Move Function reconciler to pkg/extensions
